Replace MD5 checksums with SHA256.
MD5 checksums are insecure, and there are tools that may want to consume
the index that do not support this hash algorithm.

In particular, https://github.com/bouk/arduino-nix is a project which
transforms arduino indexes into Nix derivations, however nixpkgs has
recently removed support for md5 hashes in derivations. This makes it
impossible to consume the Adafruit board index with this tool.

This was done systematically using the following script:
https://gist.github.com/plietar/301bfcd529d87fd8e091140980b8460f